added new module and necessary refactorings
