+++ kerchunk/grib2.py
@@ -3,14 +3,11 @@ import copy
 import io
 import logging
 from collections import defaultdict
-import warnings
 from typing import Iterable, List, Dict, Set, TYPE_CHECKING, Optional, Callable
 import ujson
-import itertools
 
 if TYPE_CHECKING:
     import pandas as pd
-    import datatree
 
 import fsspec
 import zarr
@@ -20,6 +17,7 @@ import numpy as np
 from kerchunk.utils import class_factory, _encode_for_JSON
 from kerchunk.codecs import GRIBCodec
 from kerchunk.combine import MultiZarrToZarr, drop
+from kerchunk._grib_idx import _map_grib_file_by_group
 
 
 try:
@@ -31,10 +29,6 @@ except ModuleNotFoundError as err:  # pragma: no cover
             "`conda install -c conda-forge cfgrib`. See https://github.com/ecmwf/cfgrib "
             "for more details."
         )
-
-
-class DynamicZarrStoreError(ValueError):
-    pass
 
 
 # cfgrib copies over certain GRIB attributes
@@ -653,236 +647,6 @@ def parse_grib_idx(
         raise ValueError(f"Attribute mapping for grib file {basename} is not unique")
 
     return result.set_index("idx")
-
-
-def build_path(path: Iterable[str | None], suffix: Optional[str] = None) -> str:
-    """
-    Returns the path to access the values in a zarr store without a leading "/"
-
-    Parameters
-    ----------
-    path : Iterable[str | None]
-        The path is the list of values to the element in zarr store
-    suffix : str
-        Last element if any
-
-    Returns
-    -------
-    str : returns the path as a string
-    """
-    return "/".join([val for val in [*path, suffix] if val is not None]).lstrip("/")
-
-
-def extract_dataset_chunk_index(
-    dset: "datatree.DataTree",
-    ref_store: Dict,
-    grib: bool = False,
-) -> list[dict]:
-    """
-    Process and extract a kerchunk index for an xarray dataset or datatree node.
-
-    The data_vars from the dataset will be indexed.
-    The coordinate vars for each dataset will be used for indexing.
-    Datatrees generated by grib_tree have some nice properties which allow a denser index.
-
-    Parameters
-    ----------
-    dset : datatree.DataTree
-        The datatree node from the datatree instance
-    ref_store : Dict
-        The zarr store dictionary backed by the gribtree
-    grib : bool
-        boolean for treating coordinates as grib levels
-
-    Returns
-    -------
-    list[dict] : returns the extracted grib metadata in the form of key-value pairs inside a list
-    """
-    import datatree
-
-    result: list[dict] = []
-    attributes = dset.attrs.copy()
-
-    dpath = None
-    if isinstance(dset, datatree.DataTree):
-        dpath = dset.path
-        walk_group = dset.parent
-        while walk_group:
-            attributes.update(walk_group.attrs)
-            walk_group = walk_group.parent
-
-    for dname, dvar in dset.data_vars.items():
-        # Get the chunk size - `chunks` property only works for xarray native
-        zarray = ujson.loads(ref_store[build_path([dpath, dname], suffix=".zarray")])
-        dchunk = zarray["chunks"]
-        dshape = dvar.shape
-
-        index_dims = {}
-        for ddim_nane, ddim_size, dchunk_size in zip(dvar.dims, dshape, dchunk):
-            if dchunk_size == 1:
-                index_dims[ddim_nane] = ddim_size
-            elif dchunk_size != ddim_size:
-                # Must be able to get a single coordinate value for each chunk to index it.
-                raise ValueError(
-                    "Can not extract chunk index for dimension %s with non singleton chunk dimensions"
-                    % ddim_nane
-                )
-            # Drop the dim where each chunk covers the whole dimension - no indexing needed!
-
-        for idx in itertools.product(*[range(v) for v in index_dims.values()]):
-            # Build an iterator over each of the single dimension chunks
-            dim_idx = {key: val for key, val in zip(index_dims.keys(), idx)}
-
-            coord_vals = {}
-            for cname, cvar in dvar.coords.items():
-                if grib:
-                    # Grib data has only one level coordinate
-                    cname = (
-                        cname
-                        if cname
-                        in ("valid_time", "time", "step", "latitude", "longitude")
-                        else "level"
-                    )
-
-                if all([dim_name in dim_idx for dim_name in cvar.dims]):
-                    coord_index = tuple([dim_idx[dim_name] for dim_name in cvar.dims])
-                    try:
-                        coord_vals[cname] = cvar.to_numpy()[coord_index]
-                    except Exception:
-                        raise DynamicZarrStoreError(
-                            f"Error reading coords for {dpath}/{dname} coord {cname} with index {coord_index}"
-                        )
-
-            whole_dim_cnt = len(dvar.dims) - len(dim_idx)
-            chunk_idx = map(str, [*idx, *[0] * whole_dim_cnt])
-            chunk_key = build_path([dpath, dname], suffix=".".join(chunk_idx))
-            chunk_ref = ref_store.get(chunk_key)
-
-            if chunk_ref is None:
-                logger.warning("Chunk not found: %s", chunk_key)
-                continue
-
-            elif isinstance(chunk_ref, list) and len(chunk_ref) == 3:
-                chunk_data = dict(
-                    uri=chunk_ref[0],
-                    offset=chunk_ref[1],
-                    length=chunk_ref[2],
-                    inline_value=None,
-                )
-            elif isinstance(chunk_ref, (bytes, str)):
-                chunk_data = dict(inline_value=chunk_ref, offset=-1, length=-1)
-            else:
-                raise ValueError(f"Key {chunk_key} has bad value '{chunk_ref}'")
-            result.append(dict(varname=dname, **attributes, **coord_vals, **chunk_data))
-
-    return result
-
-
-def extract_datatree_chunk_index(
-    dtree: "datatree.DataTree", kerchunk_store: dict, grib: bool = False
-) -> "pd.DataFrame":
-    """
-    Recursive method to iterate over the data tree and extract the data variable chunks with index metadata
-
-    Parameters
-    ----------
-    dtree : datatree.DataTree
-        The xarray datatree representation of the reference filesystem
-    kerchunk_store : dict
-        the grib_tree output for a single grib file
-    grib : bool
-        boolean for treating coordinates as grib levels
-
-    Returns
-    -------
-    pandas.Dataframe : The dataframe constructed from the grib metadata
-    """
-    import pandas as pd
-
-    result: list[dict] = []
-
-    for node in dtree.subtree:
-        if node.has_data:
-            result += extract_dataset_chunk_index(
-                node, kerchunk_store["refs"], grib=grib
-            )
-
-    return pd.DataFrame.from_records(result)
-
-
-def _map_grib_file_by_group(
-    fname: str,
-    mapper: Optional[Callable] = None,
-    storage_options: Optional[Dict] = None,
-) -> "pd.DataFrame":
-    """
-    Helper method used to read the cfgrib metadata associated with each message (group) in the grib file
-    This method does not add metadata
-
-    Parameters
-    ----------
-    fname : str
-        the file name to read with scan_grib
-    mapper : Optional[Callable]
-        the mapper if any to apply (used for hrrr subhf)
-
-    Returns
-    -------
-    pandas.Dataframe : The intermediate dataframe constructed from the grib metadata
-    """
-    import pandas as pd
-
-    mapper = (lambda x: x) if mapper is None else mapper
-    references = scan_grib(fname, storage_options=storage_options)
-
-    with warnings.catch_warnings():
-        warnings.simplefilter("ignore")
-        return pd.concat(
-            # grib idx is fortran indexed (from one not zero)
-            list(
-                filter(
-                    # filtering out the cfgrib metadata dataframe in case it is None
-                    lambda item: item is not None,
-                    [
-                        # extracting the metadata from a single message
-                        _extract_single_group(mapper(group), i, storage_options)
-                        for i, group in enumerate(references, start=1)
-                    ],
-                )
-            )
-        ).set_index("idx")
-
-
-def _extract_single_group(grib_group: dict, idx: int, storage_options: Dict):
-    import datatree
-
-    grib_tree_store = grib_tree(
-        [
-            grib_group,
-        ],
-        storage_options,
-    )
-
-    if len(grib_tree_store["refs"]) <= 1:
-        logger.info("Empty DT: %s", grib_tree_store)
-        return None
-
-    dt = datatree.open_datatree(
-        fsspec.filesystem("reference", fo=grib_tree_store).get_mapper(""),
-        engine="zarr",
-        consolidated=False,
-    )
-
-    k_ind = extract_datatree_chunk_index(dt, grib_tree_store, grib=True)
-    if k_ind.empty:
-        logger.warning("Empty Kind: %s", grib_tree_store)
-        return None
-
-    assert (
-        len(k_ind) == 1
-    ), f"expected a single variable grib group but produced: {k_ind}"
-    k_ind.loc[:, "idx"] = idx
-    return k_ind
 
 
 def build_idx_grib_mapping(
